Changed representation of grammar for better interface
Grammars are now represented as a struct in the transformer environment.
This makes the code that manipulates grammar more readable.

This has the added benefit that we can reuse the name of the grammar as the name
of the typesetter for the grammar. (still undocumented)
This provides a much nicer interface.