Abstract grammar spec and change interface to ndiff
Change the ndiff interface so I always specify language pairs, instead of
languages with respect to a single base language. This is more expressive at a
slight cost of verbosity, but also captures some use cases I'm finding common.
If a single language is given rather than a pair, display the grammar itself.

Adds a syntax-class for grammar display specs to help parse this.