Workaround a bug when a non-terminal is 'replaced'
When a nonterminal is considered replaced by the diff algorithm, the old
non-terminal would appear as a production in the definition of the new
nonterminal. This tries to detect this situation and doesn't output the replaced
non-terminal.

A better solution would be to write a grammar-diff algorithm, computing the
diffs between the non-terminal set and the production sets separately